+++ examples/pytorch/hetgnn/model.py
@@ -25,6 +25,8 @@ class HetGnnModel(BaseUnsupervisedModel):
         feature_dim: int,
         feature_type: np.dtype,
         metric=MRR(),
+        g=None,
+        sampler=None,
     ):
         """Initialize HetGnn model."""
         super(HetGnnModel, self).__init__(
@@ -33,6 +35,9 @@ class HetGnnModel(BaseUnsupervisedModel):
             feature_dim=feature_dim,
             feature_enc=None,
         )
+
+        self.g = g
+        self.sampler = sampler
 
         self.embed_d = embed_d
         self.node_type_count = node_type_count
@@ -158,6 +163,7 @@ class HetGnnModel(BaseUnsupervisedModel):
         Aggregating all the neighbors of node grouped by node types
         using BiLSTM and mean pooling.
         """
+        neigh_feats = torch.Tensor(neigh_feats)
         batch_s = int(neigh_feats.squeeze(0).shape[0] / self.neighbor_count)
 
         neigh_agg = self.content_agg(node_type, neigh_feats).view(
@@ -177,6 +183,7 @@ class HetGnnModel(BaseUnsupervisedModel):
 
         Aggregating all the heterogeneous content of node using BiLSTM and mean pooling.
         """
+        feats = torch.Tensor(feats)
         feature_list = feats.squeeze(0)  # self.features(torch.as_tensor(id_batch[0]))
         concate_embed = feature_list.view(feature_list.shape[0], 1, self.embed_d)
         concate_embed = torch.transpose(concate_embed, 0, 1)
@@ -216,6 +223,8 @@ class HetGnnModel(BaseUnsupervisedModel):
 
     def forward(self, context: dict) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:  # type: ignore[override]
         """Calculate score based on inputs in the context and return loss."""
+        if not isinstance(context, dict):
+            context = model.query(self.g, next(self.sampler))  # type: ignore # noqa
         feature_list = context["encoder"]
         inputs = context["inputs"]
         size_array = max([len(inputs[k]) for k in range(len(inputs))])