+++ examples/pytorch/hetgnn/test_args.py
@@ -13,15 +13,11 @@ from deepgnn import TrainMode, setup_default_logging_config
 from deepgnn import get_logger
 from deepgnn.pytorch.common.utils import get_python_type
 from deepgnn.pytorch.modeling import BaseModel
-from deepgnn.pytorch.common.dataset import TorchDeepGNNDataset
-from deepgnn.graph_engine import CSVNodeSampler, GraphEngineBackend
 from args import init_args  # type: ignore
 from model import HetGnnModel  # type: ignore
 from sampler import HetGnnDataSampler  # type: ignore
-from main import create_model, create_dataset, create_optimizer, init_args  # type: ignore
+from main import run_ray  # type: ignore
 from deepgnn.graph_engine.data.ppi import PPI
-from deepgnn.pytorch.common.ray_train import run_ray
-from typing import Optional
 
 
 def setup_module(module):
@@ -60,10 +56,6 @@ def test_run_args():
     # User only needs to prepare initializing function for model, dataset, optimizer and args.
     # reference: `deepgnn/pytorch/training/factory.py`
     result = run_ray(
-        init_model_fn=create_model,
-        init_dataset_fn=create_dataset,
-        init_optimizer_fn=create_optimizer,
-        init_args_fn=init_args,
         num_cpus=4,
         run_args=run_args,
     )