+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Remove our custom PyTorch trainers, use Ray instead.
 
+- Remove TorchDeepGNNDataset, deepgnn.graph_engine.samplers, deepgnn.graph_engine.prefetch and deepgnn.graph_engine.backends.
+
 - Removed `deepgnn.pytorch.common.utils.set_seed`, use ray `train.torch.enable_reproducibility(seed=0)` instead. Remove `deepgnn.pytorch.common.utils.to_cuda`.
 
 ### Fixed